requirements.txt
alembic==1.0.10
aniso8601==3.0.2
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
configparser==3.8.1
contextlib2==0.5.5
eventsourcing==7.1.6
fastdiff==0.2.0
Flask==1.0.3
Flask-GraphQL==2.0.0
flask-marshmallow==0.10.1
Flask-Migrate==2.5.0
Flask-SQLAlchemy==2.4.0
funcsigs==1.0.2
graphene==2.1.3
graphene-sqlalchemy==2.1.2
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
marshmallow==3.0.1
more-itertools==5.0.0
packaging==19.1
pathlib2==2.3.4
pluggy==0.12.0
promise==2.2.1
py==1.8.0
pycryptodome==3.8.2
pyparsing==2.4.2
pytest==4.6.5
python-dateutil==2.8.0
python-editor==1.0.4
readerwriterlock==1.0.4
requests==2.22.0
Rx==1.6.1
scandir==1.10.0
singledispatch==3.4.0.3
six==1.12.0
snapshottest==0.5.1
SQLAlchemy==1.3.3
SQLAlchemy-Utils==0.34.2
termcolor==1.1.0
typing==3.6.6
urllib3==1.25.3
wasmer==0.3.0
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.6.0
